fixed empty line

+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: jguthert <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/04/13 15:36:17 by jguthert          #+#    #+#             */
-/*   Updated: 2016/05/10 20:19:51 by jguthert         ###   ########.fr       */
+/*   Updated: 2016/05/10 21:05:29 by jguthert         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int				main(void)
 		return (1);
 	while (1)
 	{
-		print_prompt((int)nbr, g_env, l_env);
+		print_prompt(nbr, g_env, l_env);
 		if (read_init(&av) == 1)
 			continue ;
 		if (shell(av, &g_env, &l_env) == 1)
